requirements.txt
# Web framework + production server
Flask==3.0.3
gunicorn==22.0.0

# HTTP requests + session handling
requests==2.32.3

# Environment variable loader
python-dotenv==1.0.1

# Timezone + datetime helpers
pytz==2024.1
timezonefinder==6.5.4

SQLAlchemy==2.0.36
Flask-Login==0.6.3
